+++ src/main/java/com/letmeclean/dto/payment/response/PaymentDetailResponse.java
@@ -3,9 +3,11 @@ package com.letmeclean.dto.payment.response;
 import com.letmeclean.model.payment.PaymentStatus;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.Setter;
 
 import java.time.LocalDateTime;
 
+@Setter
 @Getter
 @AllArgsConstructor
 public class PaymentDetailResponse {
