+++ src/main/java/com/letmeclean/model/issuedticket/IssuedTicketRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 public interface IssuedTicketRepository extends JpaRepository<IssuedTicket, Long> {
 
-    @Query(value = "select new com.letmeclean.issuedticket.dto.response.IssuedTicketDetailResponse(m.email, t.name, t.description, it.createdAt) " +
+    @Query(value = "select new com.letmeclean.dto.issuedticket.response.IssuedTicketDetailResponse(m.email, t.name, t.description, it.createdAt) " +
             "from IssuedTicket it " +
             "join it.member m " +
             "join it.ticket t " +
